FEAT: Customized notification time
