+++ Conscious/Base.lproj/Main.storyboard
@@ -614,6 +614,9 @@
                                 <autoresizingMask key="autoresizingMask" flexibleMaxX="YES" flexibleMaxY="YES"/>
                                 <state key="normal" title="Button"/>
                                 <buttonConfiguration key="configuration" style="plain" title="設定通知"/>
+                                <connections>
+                                    <segue destination="gKA-9d-Ccn" kind="show" id="zw8-fL-U4Z"/>
+                                </connections>
                             </button>
                         </subviews>
                         <viewLayoutGuide key="safeArea" id="o3i-NA-AQa"/>
@@ -649,6 +652,40 @@
                 <placeholder placeholderIdentifier="IBFirstResponder" id="g7C-qj-6x7" userLabel="First Responder" customClass="UIResponder" sceneMemberID="firstResponder"/>
             </objects>
             <point key="canvasLocation" x="4552" y="2180"/>
+        </scene>
+        <!--設定通知-->
+        <scene sceneID="ifT-bP-oog">
+            <objects>
+                <viewController id="gKA-9d-Ccn" customClass="NotificationViewController" customModule="Conscious" customModuleProvider="target" sceneMemberID="viewController">
+                    <view key="view" contentMode="scaleToFill" id="yPR-Sq-PbT">
+                        <rect key="frame" x="0.0" y="0.0" width="393" height="852"/>
+                        <autoresizingMask key="autoresizingMask" widthSizable="YES" heightSizable="YES"/>
+                        <subviews>
+                            <datePicker contentMode="scaleToFill" fixedFrame="YES" contentHorizontalAlignment="center" contentVerticalAlignment="center" datePickerMode="time" minuteInterval="1" translatesAutoresizingMaskIntoConstraints="NO" id="Cml-To-AKu">
+                                <rect key="frame" x="0.0" y="359" width="393" height="75"/>
+                                <autoresizingMask key="autoresizingMask" widthSizable="YES" flexibleMaxY="YES"/>
+                            </datePicker>
+                            <button opaque="NO" contentMode="scaleToFill" fixedFrame="YES" contentHorizontalAlignment="center" contentVerticalAlignment="center" buttonType="system" lineBreakMode="middleTruncation" translatesAutoresizingMaskIntoConstraints="NO" id="qUY-y7-aBr">
+                                <rect key="frame" x="158" y="448" width="78" height="35"/>
+                                <autoresizingMask key="autoresizingMask" flexibleMaxX="YES" flexibleMaxY="YES"/>
+                                <state key="normal" title="Button"/>
+                                <buttonConfiguration key="configuration" style="plain" title="Submit"/>
+                                <connections>
+                                    <action selector="submitButtonTapped:" destination="gKA-9d-Ccn" eventType="touchUpInside" id="zur-xJ-eYy"/>
+                                </connections>
+                            </button>
+                        </subviews>
+                        <viewLayoutGuide key="safeArea" id="GFF-8Q-Aqi"/>
+                        <color key="backgroundColor" systemColor="systemBackgroundColor"/>
+                    </view>
+                    <navigationItem key="navigationItem" title="設定通知" id="oYP-s8-Kwo"/>
+                    <connections>
+                        <outlet property="datePicker" destination="Cml-To-AKu" id="cxu-f4-RRq"/>
+                    </connections>
+                </viewController>
+                <placeholder placeholderIdentifier="IBFirstResponder" id="8tT-dU-r8v" userLabel="First Responder" customClass="UIResponder" sceneMemberID="firstResponder"/>
+            </objects>
+            <point key="canvasLocation" x="4551.9083969465646" y="2880.9859154929577"/>
         </scene>
         <!--Enter Password View Controller-->
         <scene sceneID="n2t-Ru-QjY">
