--- Conscious/ProfilePage/View/SettingNotificationViewController.swift
@@ -1,18 +1,0 @@
-//
-//  SettingNotificationViewController.swift
-//  Conscious
-//
-//  Created by jeff on 2023/9/21.
-//
-
-import Foundation
-import UIKit
-
-class SettingNotificationViewController: UIViewController {
-
-    override func viewDidLoad() {
-        super.viewDidLoad()
-    }
-
-    @IBOu
-}
